+++ duke.txt
@@ -1,3 +1,4 @@
 [T][O] homework
 [E][O] dota  (at: 02/12/2019 1900)
 [T][X] homework
+[T][X] homework
