--- src/main/java/ByeCommand.java
@@ -1,6 +1,0 @@
-public class ByeCommand extends Command {
-
-    public String execute(TaskList currentTaskList, Ui ui, Storage storage) {
-        return ui.printBye();
-    }
-}
