--- src/main/java/Command.java
@@ -1,5 +1,0 @@
-public abstract class Command {
-
-    public abstract String execute(TaskList tasks, Ui ui, Storage storage) throws DukeException;
-
-}
