+++ src/main/java/duke/Commands/DeleteCommand.java
@@ -1,3 +1,11 @@
+package duke.Commands;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+import duke.Exceptions.DukeException;
+import duke.Parser;
+
 public class DeleteCommand extends Command {
 
     private String inputInstruction;
@@ -6,6 +14,13 @@ public class DeleteCommand extends Command {
         inputInstruction = inputString;
     }
 
+    /**
+     * Overrides execute method to achieve DeleteCommand.
+     * @param currentTaskList taskList containing all the current Task
+     * @param ui class that handles all I/O managements
+     * @param storage class that handles all file storage and loading
+     * @return a String output for the specific Commands
+     */
     public String execute(TaskList currentTaskList, Ui ui, Storage storage) {
         try {
             if (inputInstruction.length() == 6 || inputInstruction.length() == 7) {
