+++ src/main/java/duke/Commands/DoneCommand.java
@@ -1,3 +1,11 @@
+package duke.Commands;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+import duke.Exceptions.DukeException;
+import duke.Tasks.Task;
+
 public class DoneCommand extends Command {
 
     private int index;
@@ -6,6 +14,13 @@ public class DoneCommand extends Command {
         index = inputIndex;
     }
 
+    /**
+     * Overrides execute method to achieve DoneCommand.
+     * @param currentTaskList taskList containing all the current Task
+     * @param ui class that handles all I/O managements
+     * @param storage class that handles all file storage and loading
+     * @return a String output for the specific Commands
+     */
     public String execute(TaskList currentTaskList, Ui ui, Storage storage) {
         try {
             if (currentTaskList.getNoOfTask() < index || index <= 0) {
