+++ src/main/java/duke/Commands/InvalidCommand.java
@@ -1,0 +1,20 @@
+package duke.Commands;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+import duke.Exceptions.DukeException;
+
+public class InvalidCommand extends Command {
+
+    /**
+     * Overrides execute method to achieve InvalidCommand.
+     * @param currentTaskList taskList containing all the current Task
+     * @param ui class that handles all I/O managements
+     * @param storage class that handles all file storage and loading
+     * @return a String output for the specific Commands
+     */
+    public String execute(TaskList currentTaskList, Ui ui, Storage storage) {
+        return ui.printException(new DukeException("invalid"));
+    }
+}
