+++ src/main/java/duke/TaskList.java
@@ -1,4 +1,6 @@
-import java.util.AbstractCollection;
+package duke;
+import duke.Tasks.Task;
+import duke.Exceptions.DukeException;
 import java.util.ArrayList;
