+++ src/main/java/duke/Tasks/Deadline.java
@@ -1,3 +1,5 @@
+package duke.Tasks;
+
 import java.text.ParseException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
